Fix - Moved Add Customer button to header section instead of payment section
